requirements.txt
requests
beautifulsoup4
lxml
tqdm
pypdf
pdfminer.six
nltk
python-docx
streamlit
sentence-transformers
rapidfuzz
arxiv
scholarly